Reorganize project and add test
Give a Julia project structure
