+++ OpenMPThreading/src/main.jl
@@ -1,0 +1,9 @@
+
+import XSBenchOpenMPThreading
+
+
+function main(args)
+    XSBenchOpenMPThreading.mainImpl(args)
+end
+
+main(ARGS)
