+++ OpenMPThreading/test/runtests.jl
@@ -1,0 +1,8 @@
+
+using Test
+
+import XSBenchOpenMPThreading
+
+@testset "io" begin
+    include("test_io.jl")
+end
